pyproject.toml
[tool.poetry]
name = "14.08.22sendemail"
version = "0.1.0"
description = ""
authors = ["Ariulis <andrey59995@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
python-dotenv = "^0.20.0"
loguru = "^0.6.0"
Jinja2 = "^3.1.2"
WTForms = "^3.0.1"
python-multipart = "^0.0.5"
aiosmtplib = "^1.1.6"

[tool.poetry.dev-dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
